Testing single precision

+++ OASIMlib/common.f90
@@ -4,5 +4,5 @@ module oasim_common
     implicit none
     
     integer, parameter :: string_length = 1024
-    integer, parameter :: real_kind = c_double
+    integer, parameter :: real_kind = c_float
 end module oasim_common
